Fix MSVC compiler warnings
reg.: conversion from 'size_t' to 'muse::mpe::staff_layer_idx_t', possible loss of data (C4267)

+++ src/instrumentsscene/view/instrumentsonscorelistmodel.cpp
@@ -200,7 +200,7 @@ int InstrumentsOnScoreListModel::resolveInstrumentSequenceNumber(const String& i
     }
 
     const InstrumentTemplateList& allTemplates = repository()->instrumentTemplates();
-    return allTemplates.size();
+    return static_cast<int>(allTemplates.size());
 }
 
 void InstrumentsOnScoreListModel::addInstruments(const QStringList& instrumentIdList)
